Transformation: move all sandbox-related code into /pipeline-execution/sandbox package
