+++ transformation/src/interfaces/jobResult.ts
@@ -1,5 +1,5 @@
 import Stats from './stats'
-import JobError from './jobError'
+import JobError from '@/pipeline-execution/sandbox/jobError'
 
 export default interface JobResult {
   data?: object;
