+++ transformation/src/jsTransformationService.ts
@@ -1,11 +1,11 @@
 import axios from 'axios'
 
-import ExecutionResult from './interfaces/executionResult'
+import ExecutionResult from '@/pipeline-execution/sandbox/executionResult'
 import JobResult from './interfaces/jobResult'
 import Stats from './interfaces/stats'
 
 import TransformationService from './interfaces/transformationService'
-import SandboxExecutor from './interfaces/sandboxExecutor'
+import SandboxExecutor from '@/pipeline-execution/sandbox/sandboxExecutor'
 
 const VERSION = '0.0.2'
