+++ transformation/src/pipeline-execution/sandbox/executionResult.ts
