+++ transformation/src/pipeline-execution/sandbox/sandboxExecutor.ts
